% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimerge.R
\name{multimerge}
\alias{multimerge}
\title{Merge multiple NARS datasets by SITE_ID and VISIT_NO}
\usage{
multimerge(mergeme)
}
\arguments{
\item{mergeme}{A folder where processed datasets are stored. Typically would select
the variables of interest and save the modified dataset in a separate folder with other processed
datasets to merge together}
}
\value{
A data frame with the merged datasets joined based on SITE_ID and VISIT_NO
}
\description{
Merge multiple NARS datasets by SITE_ID and VISIT_NO
}
\examples{
nrsa<- multimerge("processed_data") # name of folder with processed data to merge
}
