% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_calc.R
\name{R2_calc}
\alias{R2_calc}
\title{Process SEM R2 for point figure for manuscript 6/2/2023}
\usage{
R2_calc(R2out_merge)
}
\arguments{
\item{R2out_merge}{Merged observed R2 and max R2 output dataframes to modify}
}
\value{
Processed dataframe with percentage of maximum R2 that was explained in the model and relabled variables
}
\description{
Process SEM R2 for point figure for manuscript 6/2/2023
}
\examples{
R2_proc_all <- R2_calc(merged_R2_df) # Apply function on merged dataframe to calculate \%max R2 explained in the model
}
